requirements.txt
numpy
pandas
pymysql
python-dotenv
scikit-learn
seaborn
xgboost
catboost